requirements.txt
openai~=0.27.4
streamlit~=1.21.0
python-dotenv~=1.0.0